+++ zcash_client_backend/src/sync.rs
@@ -28,8 +28,8 @@ use zcash_primitives::{
 use crate::{
     data_api::{
         chain::{
-            error::Error as ChainError, scan_cached_blocks, BlockCache, ChainState,
-            CommitmentTreeRoot,
+            error::Error as ChainError, scan_cached_blocks, truncate_block_cache, BlockCache,
+            ChainState, CommitmentTreeRoot,
         },
         scanning::{ScanPriority, ScanRange},
         WalletCommitmentTrees, WalletRead, WalletWrite,
@@ -60,7 +60,7 @@ pub async fn run<P, ChT, CaT, DbT>(
     batch_size: u32,
 ) -> Result<(), Error<CaT::Error, <DbT as WalletRead>::Error, <DbT as WalletCommitmentTrees>::Error>>
 where
-    P: Parameters + Send + 'static,
+    P: Parameters + Send + Sync + 'static,
     ChT: GrpcService<BoxBody>,
     ChT::Error: Into<StdError>,
     ChT::ResponseBody: Body<Data = Bytes> + Send + 'static,
@@ -68,7 +68,7 @@ where
     CaT: BlockCache,
     CaT::Error: std::error::Error + Send + Sync + 'static,
     DbT: WalletWrite + WalletCommitmentTrees,
-    DbT::AccountId: ConditionallySelectable + Default + Send + 'static,
+    DbT::AccountId: ConditionallySelectable + Default + Send + Sync + 'static,
     <DbT as WalletRead>::Error: std::error::Error + Send + Sync + 'static,
     <DbT as WalletCommitmentTrees>::Error: std::error::Error + Send + Sync + 'static,
 {
@@ -106,7 +106,7 @@ async fn running<P, ChT, CaT, DbT, TrErr>(
     #[cfg(feature = "transparent-inputs")] wallet_birthday: BlockHeight,
 ) -> Result<bool, Error<CaT::Error, <DbT as WalletRead>::Error, TrErr>>
 where
-    P: Parameters + Send + 'static,
+    P: Parameters + Send + Sync + 'static,
     ChT: GrpcService<BoxBody>,
     ChT::Error: Into<StdError>,
     ChT::ResponseBody: Body<Data = Bytes> + Send + 'static,
@@ -114,7 +114,7 @@ where
     CaT: BlockCache,
     CaT::Error: std::error::Error + Send + Sync + 'static,
     DbT: WalletWrite,
-    DbT::AccountId: ConditionallySelectable + Default + Send + 'static,
+    DbT::AccountId: ConditionallySelectable + Default + Send + Sync + 'static,
     DbT::Error: std::error::Error + Send + Sync + 'static,
 {
     // 3) Download chain tip metadata from lightwalletd
@@ -403,22 +403,16 @@ async fn scan_blocks<P, CaT, DbT, TrErr>(
     scan_range: &ScanRange,
 ) -> Result<bool, Error<CaT::Error, <DbT as WalletRead>::Error, TrErr>>
 where
-    P: Parameters + Send + 'static,
+    P: Parameters + Send + Sync + 'static,
     CaT: BlockCache,
     CaT::Error: std::error::Error + Send + Sync + 'static,
     DbT: WalletWrite,
-    DbT::AccountId: ConditionallySelectable + Default + Send + 'static,
+    DbT::AccountId: ConditionallySelectable + Default + Send + Sync + 'static,
     DbT::Error: std::error::Error + Send + Sync + 'static,
 {
     info!("Scanning {}", scan_range);
-    let scan_result = scan_cached_blocks(
-        params,
-        db_cache,
-        db_data,
-        scan_range.block_range().start,
-        initial_chain_state,
-        scan_range.len(),
-    );
+    let scan_result =
+        scan_cached_blocks(params, db_cache, db_data, initial_chain_state, scan_range).await;
 
     match scan_result {
         Err(ChainError::Scan(err)) if err.is_continuity_error() => {
@@ -443,8 +437,7 @@ where
             // This does imply that assumed-valid blocks will be re-downloaded, but it is
             // also possible that in the intervening time, a chain reorg has occurred that
             // orphaned some of those blocks.
-            db_cache
-                .truncate(rewind_height)
+            truncate_block_cache(db_cache, rewind_height)
                 .await
                 .map_err(Error::Cache)?;